refactor(framework): split resource_routes.py into per-verb mounts/ package
The 2069-line `framework/dispatch/resource_routes.py` is the largest
file in the repo and the central piece of the spec→route pipeline.
Splitting it lets each mount fit on a screen and makes per-verb
history readable, without churning any external caller.

What moved:
- `src/framework/dispatch/mounts/{create,delete,detail,form,list_,
  update,state_axis,related_list,search,edge,entity}.py` — one
  `mount_*` function per file. `mount_entity` (the spec-driven
  orchestrator) lives in `entity.py` alongside the helper
  `_detect_caller_module` (kept here so the stack-walk continues to
  skip frames inside its own module).
- `mounts/_spec.py` — the `QueryParam` and `ResourceSpec`
  dataclasses + the `_UNSET` sentinel.
- `mounts/_synth.py` — the route-signature synthesis machinery
  (`synthesize_route_fn`, `SynthOptions`, `MountError`,
  `is_optional` / `is_pydantic_model`).
- `mounts/_common.py` — small helpers shared across the per-verb
  modules (path building, parent walks, dotted-path resolution,
  the `resolve_handler` test-patch indirection, normalize_filters,
  owned_factory_makers).

`resource_routes.py` becomes a thin re-export shim that lifts the
public surface (`mount_entity`, every `mount_*`,
`ResourceSpec`, `QueryParam`, `MountError`) plus the private
`_wrap_state_axis_with_self_guard` (patched by contract tests) and
the `APIResponse` class (mutated by unit-test monkeypatches) out of
the new package. Every external import path keeps working.

`mount_entity` looks up the per-verb mount functions through the
shim at call time rather than capturing them at import time. That
preserves the long-standing test pattern of monkeypatching
`resource_routes.mount_list = stub` to verify dispatch — patches
against the shim's namespace are still honored on each call.

`src/framework/dispatch/README.md` now points at the new
`mounts/README.md` for the per-file layout.
