+++ src/framework/dispatch/README.md
@@ -24,7 +24,8 @@ mount's query_params= tuple).
 ## Files
 
 - `entity_spec.py` — `EntitySpec` dataclass + supporting types (`StateAxis`, `RouteSet`, `Templates`, `EdgeAudit`, `AuthDeps`, `AuthzPolicy`, `M2NRelation`, `RelatedListSubresource`, `Redirects`). The spec is the single declaration the rest of the codebase reads from. `to_resource_spec()` bridges to the mount helpers.
-- `resource_routes.py` — the `mount_entity` dispatcher and the underlying `mount_*` family (`mount_list`, `mount_detail`, `mount_create`, `mount_update`, `mount_delete`, `mount_form`, `mount_state_axis`, `mount_related_list`, `mount_edge_routes`). Plus the `ResourceSpec` dataclass `mount_entity` constructs internally from the upstream `EntitySpec`. Per-mount docstrings document required spec fields and handler signatures.
+- `mounts/` — the `mount_entity` dispatcher and the per-verb `mount_*` family, one file per verb. See [`mounts/README.md`](mounts/README.md) for the file-by-file layout. The `ResourceSpec` dataclass `mount_entity` constructs internally from the upstream `EntitySpec` lives at `mounts/_spec.py`. Per-mount docstrings document required spec fields and handler signatures.
+- `resource_routes.py` — re-export shim that lifts the public surface (`mount_entity`, the `mount_*` family, `ResourceSpec`, `QueryParam`, `MountError`) out of `mounts/` so external imports of `src.framework.dispatch.resource_routes` keep working.
 - `handlers.py` — generic CRUD handlers (`handle_create`, `handle_update`, `handle_delete`, `handle_detail`, `handle_list`) and the `make_<verb>_handler(spec)` factories that build callables with synthesized signatures so `mount_*`'s introspection wires the right deps.
 - `filters.py` — declarative filter types (`Filter` base, `TextFilter`, `ChoiceFilter`) that an entity declares on `EntitySpec.filters`. Each carries both the URL contract (`name`, `annotation`, `default` — bridged to `QueryParam` via `to_query_param()` so FastAPI sees a normal `Query(...)` param) and UI metadata (`kind`, `label`, `placeholder`, `choices`, `multi`). Every declared `Filter` renders as a form control on the dedicated `/<collection>/search` page (`views/search.html`); the list-page toolbar carries only the "Filter · N" link (left) and the page-action menu (right). Active values appear as removable tags in the active-filter strip below the toolbar. Raw `QueryParam` entries on `EntitySpec.filters` still work (URL-only, no UI); `Filter` is layered on top, not a replacement.
 - `pagination.py` — `Page` snapshot dataclass + `parse_page` / `offset_for` / `paginate` / `base_query` helpers consumed by `handle_list` and the bespoke list handlers (`handle_list_my_favorites`, `handle_list_user_providers`). Reads `?page=N` from the request, asks the logic layer for `per_page + 1` rows, slices the probe off to compute `has_next` without a `COUNT(*)`. `DEFAULT_PAGE_SIZE = 25`; per-entity override via `EntitySpec.page_size`. The view-type template `views/list.html` renders the `_shared/pagination.html` footer automatically from the `page_meta` context var; pages where the result fits on a single page emit nothing.
@@ -65,7 +66,7 @@ handler would silently bypass the spec hook. Use one or the other.
 
 ## Handler stitching
 
-`mount_entity` walks up the call stack to find the route module that invoked it and `setattr`s factory-built handlers onto it as `_handle_<verb>_<entity>`. This keeps the contract-test patch path stable across the refactor that moved handler bodies out of route files. The stitching is described in the docstring of `_detect_caller_module` in `resource_routes.py`.
+`mount_entity` walks up the call stack to find the route module that invoked it and `setattr`s factory-built handlers onto it as `_handle_<verb>_<entity>`. This keeps the contract-test patch path stable across the refactor that moved handler bodies out of route files. The stitching is described in the docstring of `_detect_caller_module` in `mounts/entity.py`.
 
 ## Tests
