+++ src/framework/dispatch/mounts/create.py
@@ -1,0 +1,91 @@
+"""`mount_create` — `POST /<collection>`."""
+
+from typing import Any, Awaitable, Callable
+
+from fastapi import Request, status
+
+from src.framework.dispatch.mounts._common import (
+    call_handler_with,
+    parent_path_param_pairs,
+    path_segments_under_router,
+    walk_parent_chain,
+)
+from src.framework.dispatch.mounts._spec import ResourceSpec
+from src.framework.dispatch.mounts._synth import SynthOptions, synthesize_route_fn
+from src.framework.http.forms import parse_and_validate_form
+from src.framework.http.responses import created_response
+
+
+def mount_create(
+    router: Any,
+    spec: ResourceSpec,
+    handler: Callable[..., Awaitable[Any]],
+) -> None:
+    """Mount ``POST /<collection>``.
+
+    Parses a form-encoded body via ``spec.create_adapter`` (a Pydantic
+    ``TypeAdapter``) and calls the handler with ``payload=``, plus any
+    typed deps the handler declares (``repo``, ``audit_repo``,
+    ``requesting_user``, any extra repos via the registry). The handler
+    is expected to use ``mutate(verb="create")`` so the audit row +
+    commit are owned by the context manager.
+
+    Response is ``201 Created`` with ``Location`` and ``HX-Redirect`` set.
+    Defaults: ``Location: /<collection>/<new_id>``, ``HX-Redirect`` =
+    ``spec.create_redirect(...)`` if set, else ``Location``. The
+    ``create_redirect`` callable receives the new id under ``spec.id_param``
+    so it can build a per-resource target (e.g. providers redirect to
+    ``/clinicians/{id}/form`` after create).
+
+    Requires: ``spec.write_user_dep``, ``spec.create_adapter``.
+    """
+    if spec.write_user_dep is None:
+        raise ValueError(
+            f"mount_create requires {spec.collection!r} to set write_user_dep."
+        )
+    if spec.create_adapter is None:
+        raise ValueError(
+            f"mount_create requires {spec.collection!r} to set create_adapter."
+        )
+
+    collection = spec.collection
+    id_param = spec.id_param
+    create_adapter = spec.create_adapter
+    create_redirect = spec.create_redirect
+
+    path = path_segments_under_router(spec, with_id=False)
+    parent_id_names = tuple(p[0] for p in parent_path_param_pairs(spec))
+
+    async def response_builder(*, handler, handler_kwarg_names, kwargs):
+        request: Request = kwargs["request"]
+        kwargs["payload"] = await parse_and_validate_form(request, create_adapter)
+        created = await call_handler_with(handler, handler_kwarg_names, kwargs)
+        path_kwargs = {name: kwargs[name] for name in parent_id_names}
+        # Default Location is the canonical resource URL — for a top-level
+        # resource that's /<collection>/{id}; for a sub-resource we point
+        # at the parent because the spec doesn't have a "list children"
+        # canonical URL convention. Per-resource overrides via
+        # `create_redirect` handle the HX-Redirect target.
+        if spec.parent is None:
+            location = f"/{collection}/{created.id}"
+        else:
+            top = walk_parent_chain(spec)[0]
+            top_id = path_kwargs[top.id_param]
+            location = f"/{top.collection}/{top_id}"
+        if create_redirect is not None:
+            hx = create_redirect(**path_kwargs, **{id_param: created.id})
+        else:
+            hx = location
+        return created_response(id=created.id, location=location, hx_redirect=hx)
+
+    route_fn = synthesize_route_fn(
+        handler=handler,
+        spec=spec,
+        options=SynthOptions(
+            user_dep=spec.write_user_dep,
+            body_adapter=create_adapter,
+            path_param_names=parent_id_names,
+        ),
+        response_builder=response_builder,
+    )
+    router.post(path, status_code=status.HTTP_201_CREATED)(route_fn)
